requirements.txt
Python 2.7 or 3.6
Flask
Flask-Login
Flask-SocketIO
Flask-WTForms
SQLAlchemy
SQLite3
Gevent